# Initial version of an adjacency matrix based graph