This version adds a `DenseGraph` struct, which implements the basic
graph functionality. The underlying structure is two weight maps and an
adjacency matrix.

The following traits are currently implemented:
- Default
- Index/IndexMut
- Clone

As well as basic node/edge addition.

The DenseMapping/IdGenerator structs allow fast indexing of weights,
combined with automatic index generation. This comes at the cost of
manual and regular resizing of a slice (`std::vec::Vec`). This slice can
probably be replaced with a `std::collections::HashMap` (since it uses
open-addressing under the hood), [stash](https://github.com/Stebalien/stash-rs) or even [vec_map](https://github.com/contain-rs/vec-map).

I was largely inspired from [src/graph.rs](https://github.com/bluss/petgraph/blob/master/src/graph.rs) for most of the implementation, so there's probably a bunch of documentation errors to fix.

__Important note:__ I'm aware there's still a bunch of work to be done and I'll probably need some pointers (#28) concerning the methodology for implementing neighbor iterators/visiting.