@jmcomets 2017-01-07
For reference, I wrote an implementation using stash, available [here](https://github.com/jmcomets/petgraph/commit/9d5c2f871f302c1212907eff0924d8f62de8c872). It's significantly shorter and much easier to understand, although it does imply adding the stash dependency. I also suspect it's faster by looking at stash's source code.