@bluss 2017-01-07
Thanks will look at this later. Don't add a stash dep. I suspect it's about the same as StableGraph's freelist thing