@jmcomets 2017-02-25
I've just squashed everything, could you have a look? I'm planning to add a bunch of unit-tests, but the implementation should be fairly complete.