@jmcomets 2017-03-06
Yeah I was expecting this...

In these use-cases, I like to grow vecs to the next power of two, instead
of simply growing to "len + 1". Thanks to this, the vec grows
exponentially, therefore performing much less allocations.

On 19:29, Mon, Mar 6, 2017 bluss <notifications@github.com> wrote:

> *@bluss* commented on this pull request.
> ------------------------------
>
> In src/dense_graph.rs
> <https://github.com/bluss/petgraph/pull/120#discussion_r104485888>:
>
> > +    IntoNeighbors,
> +    IntoNeighborsDirected,
> +    IntoNodeIdentifiers,
> +    NodeCount,
> +    Visitable
> +};
> +
> +pub use graph::{node_index, edge_index};
> +
> +type OptionVec<T> = Vec<Option<T>>;
> +
> +const INITIAL_RESIZE: usize = 10;
> +
> +fn ensure_len<T>(vec: &mut OptionVec<T>, min_len: usize) where T: Clone {
> +    if vec.len() <= min_len {
> +        let next_power_of_2_len = (min_len + 2) & !1;
>
> I don't understand how this works.
>
> I would probably elect to let the vec decide pace to grow, and just ask it
> to grow to the min_len, and fill it with None until its resulting capacity.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/bluss/petgraph/pull/120#pullrequestreview-25325990>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABzmg3QO4YL0AdD1WAnTvPeyIdsiQ9Oiks5rjFB0gaJpZM4LdbdD>
> .
>