@bluss 2017-03-06
That's the default allocation pattern for a vec, we don't need to reimplement that though.