@jmcomets 2017-03-06
I've just run some benchmarks, and while looking at it, I noticed that I screwed up when writing this. It simply goes to the next *multiple* of 2, not the next power...

Here are my results:
```rust
// with:
// test dense_graph_add_edges_from_root ... bench:      15,340 ns/iter (+/- 988)
//
// without:
// test dense_graph_add_edges_from_root ... bench:      14,963 ns/iter (+/- 4,461)
#[bench]
fn dense_graph_add_edges_from_root(bench: &mut Bencher)
{
    bench.iter(|| {
        let mut gr: DenseGraph<(), ()> = DenseGraph::new();
        let a = gr.add_node(());

        for _ in 0..100 {
            let b = gr.add_node(());
            gr.add_edge(a, b, ());
        }   
    }); 
}

// with:
// test dense_graph_add_adjacent_edges ... bench:      40,118 ns/iter (+/- 26,898)
//
// without:
// test dense_graph_add_adjacent_edges ... bench:      34,736 ns/iter (+/- 28,208)
#[bench]
fn dense_graph_add_adjacent_edges(bench: &mut Bencher)
{
    bench.iter(|| {
        let mut gr: DenseGraph<(), ()> = DenseGraph::new();
        let mut prev: Option<NodeIndex> = None;
        for _ in 0..100 {
            let b = gr.add_node(());

            if let Some(a) = prev {
                gr.add_edge(a, b, ());
            }   

            prev = Some(b);
        }   
    }); 
}
```

And here's the patch:
```diff
@@ -41,8 +41,8 @@ const INITIAL_RESIZE: usize = 10;

fn ensure_len<T>(vec: &mut OptionVec<T>, min_len: usize) where T: Clone {
     if vec.len() <= min_len {
-        let next_power_of_2_len = (min_len + 2) & !1;
-        vec.resize(cmp::max(next_power_of_2_len, INITIAL_RESIZE), None);
+        let next_len = min_len + 1;
+        vec.resize(cmp::max(next_len, INITIAL_RESIZE), None);
     }
 }
```


I'll attach the benchmarks and the patch to this in a follow-up commit.