@bluss 2017-03-06
Well that explains why I didn't understand that computation.