@jmcomets 2017-10-07
I'm going to close this PR. I'm not satisfied with the benchs I've run on it. I have another version which is almost ready in [`adjacency-matrix-v2`](https://github.com/jmcomets/petgraph/tree/adjacency-matrix-v2), which uses more cache-friendly structures.